backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filetype==1.0.10
idna==3.3
PyJWT==2.6.0
pytz==2022.6
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.3
ua-parser==0.10.0
urllib3==1.26.8
websocket-client==0.59.0